requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
pillow==11.3.0
numpy==2.1.1
ultralytics==8.3.78
qdrant-client==1.15.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
torchvision==0.23.0+cpu
open-clip-torch==2.32.0
pytest==8.4.1
httpx==0.28.1
